app/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core==0.3.75
langchain-google-genai
langgraph==0.6.6
langsmith
python-dotenv
pandas==2.3.2
sqlparse==0.5.3
pydantic==2.11.7

annotated-types
anyio
certifi
charset-normalizer
colorama
distro
h11
httpcore
httpx
idna
jiter
jsonpatch
jsonpointer
langchain-openai
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
numpy
openai
orjson
ormsgpack
packaging
pydantic_core
python-dateutil
pytz
PyYAML
regex
requests
requests-toolbelt
six
sniffio
tenacity
tiktoken
tqdm
typing-inspection
typing_extensions
tzdata
urllib3
xxhash
zstandard